requirements.txt
fastapi
uvicorn
jinja2
python-multipart
speechrecognition
pydub
sqlalchemy
passlib[bcrypt]
PyJWT
pymysql